requirements.txt
requests
pytyhon-dotenv
FastAPI
uvicorn
requests
pydantic